requirements.txt
image
pillow
pyserial
